Added only_connect.py and updated app.py
app.py has been updated to utilize more user-defined class objects, decluttering the main code page and placing the bulk of the work inside of only_connect.py. only_connect.py defines class objects and functions to aid in the running of the game. Still working on developing the timer for use in the first three rounds.